dev_environment/workspace/requirements.txt
# suggested packages - remove/add based on your needs
python-dotenv
pyautogen
openai